// src/zorkin-client/zorkin-server-oapi-client/api.ts
export interface AuthorizeSessionRequest {
    /**
     * 
     * @type {string}
     * @memberof AuthorizeSessionRequest
     */
    'clientJWT': string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizeSessionRequest
     */
    'sessionPK': string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizeSessionRequest
     */
    'clientID': string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizeSessionRequest
     */
    'groupTxnIdSessionSignature': string;
}
// ...
export interface GetAuthorizeSessionGroupHashRequest {
    /**
     * 
     * @type {string}
     * @memberof GetAuthorizeSessionGroupHashRequest
     */
    'clientJWT': string;
    /**
     * 
     * @type {string}
     * @memberof GetAuthorizeSessionGroupHashRequest
     */
    'sessionPK': string;
    /**
     * 
     * @type {string}
     * @memberof GetAuthorizeSessionGroupHashRequest
     */
    'clientID': string;
}
// ...
export interface GetContractAccountRequest {
    /**
     * 
     * @type {string}
     * @memberof GetContractAccountRequest
     */
    'clientJWT': string;
    /**
     * 
     * @type {string}
     * @memberof GetContractAccountRequest
     */
    'clientID': string;
}